refactoring download code into entity
