+++ cartoframes/data/observatory/geography.py
@@ -1,18 +1,8 @@
 from __future__ import absolute_import
 
-from warnings import warn
-
-from google.api_core.exceptions import NotFound
-
-from carto.exceptions import CartoException
-
 from .entity import CatalogEntity
 from .repository.dataset_repo import get_dataset_repo
 from .repository.geography_repo import get_geography_repo
-from ..clients.bigquery_client import BigQueryClient
-from ...auth import get_default_credentials
-
-_WORKING_PROJECT = 'carto-do-customers'
 
 
 class Geography(CatalogEntity):
@@ -64,27 +54,4 @@ class Geography(CatalogEntity):
         return self.data['summary_jsonb']
 
     def download(self, credentials=None):
-        credentials = _get_credentials(credentials)
-        user_dataset = credentials.username.replace('-', '_')
-        bq_client = _get_bigquery_client(_WORKING_PROJECT, credentials)
-
-        project, dataset, table = self.id.split('.')
-        view = 'view_{}_{}'.format(dataset.replace('-', '_'), table)
-
-        try:
-            file_path = bq_client.download_to_file(_WORKING_PROJECT, user_dataset, view)
-        except NotFound:
-            raise CartoException('You have not purchased the dataset `{}` yet'.format(self.id))
-
-        warn('Data saved: {}.'.format(file_path))
-        warn("Read it by: `pandas.read_csv('{}')`.".format(file_path))
-
-        return file_path
-
-
-def _get_credentials(credentials=None):
-    return credentials or get_default_credentials()
-
-
-def _get_bigquery_client(project, credentials):
-    return BigQueryClient(project, credentials)
+        return self._download(credentials)
